Restructure CLI into separate command classes
Refactor the monolithic CLI class into separate command classes using
WP-CLI's nested command pattern for better organization and maintainability.

- Split includes/class-cli.php into includes/cli/ directory
- Create separate command classes: Post, Comment, Actor, Outbox, Move, Follow, Self_Destruct
- Add Command base class with version subcommand
- Rename self_destruct to self-destruct (underscore to hyphen)
- Group undo/reschedule under outbox subcommand
- Keep class-cli.php as registry that registers all subcommands
